+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/extensions/options/EnableableOptions.kt
@@ -6,3 +6,12 @@ public interface EnableableOptions {
 
     public var Project.isEnabled: Boolean
 }
+
+public interface EnableableAllOptions {
+
+    public var Project.allEnabled: Boolean
+
+    public fun Project.enableAll(enable: Boolean = true) {
+        allEnabled = enable
+    }
+}